adding Owner Dashboard: - Device List View - Device Details View

+++ database/migrations/device/2019_11_26_113614_create_table_chipsets.php
@@ -24,7 +24,7 @@ class CreateTableChipsets extends Migration
             $table->timestamps();
             $table->bigIncrements('id');
 			$table->string("name", 100)->default("");
-			$table->unsignedSmallInteger("used_io")->default(0);
+			$table->unsignedTinyInteger("used_io")->default(0);
         });
     }
